Create MultipleCIToTask.js
Endpoint:

https://<<<YOUR_INSTANCE_NAME>>>.service-now.com/api/gmi/related_live_ci_to_task

Payload Details:

User posts a JSON object with the task number and an array of CIs to associate to that task. Methods to identify the CIs being inserted can be sys_id, serial, FQDN, name, and IP, with that order of preference. Typically, we will only expect 1 identifier per CI line. fields in the post body:

Task: Task number that we should associate the CIs to. Required field, fail and return error code if this is not populated.

query_string: optional field that allows the user to pass in an override query string that we will add while querying for CIs. could allow for teams to only find CIs owned by a certain team, active CIs, retired CIs, only CIs of certain classes etc. if this is not provided or is an empty string, by default we will only search for live systems.

cis: Array of CIs that should be associated to the provided task. In each CI, can provide one of several methods to identify the CI, including name, serial, sys_id, IP, or FQDN. This same logic can be extended to include more custom fields if required.

I am using u_cmdb_ci_status for verifying production CI. it can be OOB field ( install_status) of whatever the customer is using.