Update CONTRIBUTING.md
Issue: Github does not automatically switch you over the new branch when it is created. 

Including clarification steps to prevent users from committing updates to the main branch. We ask them to create a new branch, but never instruct them to switch over to that branch. As per Earl's feedback - "This can be averted in the future just by making sure you always create a new branch when working on something new. And specifically, creating a new branch from a clean main branch."