Business rule (#712)
* Create Replace Script.js

This Script help to update the author of knowledge article, if author leaves the organization.

* Create Readme.md

This business rule help to replace the KB author of knowledge article to Knowledge base manager, if the author of that article leaves the organization. This is a kind of logic which helps to assign the knowledge article to knowledge base manager.