Create Display message if both Peer reviewer and requestor  are same person
This business rule is used to display a warning message in the change management module when a request by and peer reviewer is the same person. Below is the error message display,

"The Requester and Peer Reviewer cannot be the same person.  The change request has been sent to the Peer Reviewer group specified and someone else from your team will need to perform the peer review"