Display warningmessage (#824)
* Create Display message if both Peer reviewer and requestor  are same person

This business rule is used to display a warning message in the change management module when a request by and peer reviewer is the same person. Below is the error message display,

"The Requester and Peer Reviewer cannot be the same person.  The change request has been sent to the Peer Reviewer group specified and someone else from your team will need to perform the peer review"

* Create readme.md

This Business rule is for displaying warning messages when requested by and peer reviewer are the same person.

* Warnmessagescript.jss

This business rule should be run "before" on Insert.

* Create Warningmessage_Business rule.jss

* Delete Business Rules/Display warning message when peer reviewer and Requested by are same personWarnmessagescript.jss

* Delete Business Rules/Display message if both Peer reviewer and requestor  are same person

* Update readme.md

* Update readme.md