Get field from glide record (#927)
* Create getField.js

Added a code for function to get the fields array from the Glide record

* Create readme.md

The code snippet provided is a JavaScript function that retrieves and logs the field names from a specified GlideRecord. The function getFields takes a GlideRecord instance as a parameter and returns an array containing the names of all the fields in the specified record.