Create Send reminders weekly
Consider the scenario of sending remainder approvals weekly. 

Here is the code on the scheduled job that runs daily but sends notification after every week(7 days) based on the due date.

Same code can be applied to any other table, where you to trigger an event or edit a record on weekly basis based on a certain Date field.