PreventFutureDateSelection
Table: Select the table where the date field is located (e.g., Incident).
Applies: Specify the date field you want to validate (e.g., opened_at).

when a user tries to select a future date in the specified date field, the script will prevent it and display an alert. You should replace 'your_date_field' with the actual name of the date field in your instance.