Bookmarklet create story task (#995)
* Refactored the impersonation dialog

Moved the impersonation dialog from a new tab into a popup window. The instance page is refreshed with the new user's context, but the location remains the same.

* Create Story Task

When viewing a story record, this bookmarklet will create a new child task record and pre-set defined values.

* Delete Browser Bookmarklets/Impersonation/readme.md

This file was captured by mistake and is included in another branch.

* Delete Browser Bookmarklets/Impersonation/impersonation.js

This file was captured by mistake and is included in a different branch.