Create get_set_mrvs_variable_value.js
How to fetch MRVS value using g_service_catalog

Use this onLoad catalog client script to fetch the MRVS variable value and hide the variable.  In this example, a catalog item for blocking multiple IP addresses on a firewall has a variable **address_type** with two choices - IPV4 and IPV6. The MRVS has two variables (ipv4_address and ipv6_address) for the respective address types. If the Address type field on the parent form is set to IPV4, the field IPV6 address is hidden on the MRVS.