Create integrationBTW2IntsancesWithIncidentNumber.js
Though It is my first time to participate in the HactoberFest, I believe that I put my 100% effort and this is my 4th pull request. and I am happy to contribute more.

This code helps in integrating two ServiceNow instances using Rest Message - in that HTTP Post method, where I also included in the code that, once an incident is created in the source instance same record will be created in the target instance and the target record's sys_id and incident number will be auto-populated in the source instance's incident record's Correlation ID& Display field's respectively.