Create readme.md
Using UI action script we can create a Problem task from a Problem and associate it to the current problem

Short description of the problem task is "Problem Task Created for problem " + current.number
Same will be added to the Description as well.
sys_id of the current problem record will be added to the problem field of the Problem Task. This will create relation between Problem and Problem Task.
Problem Task Type will be General

UI Action will be available as Form Button on the Problem Form.
When Clicked Problem Task will be Created and associated to the current Problem.