Inactive user cleanup (#1166)
* Create InactiveUserCleanup.js

This Script Include, InactiveUserCleanup, has a method deactivateInactiveUsers that takes the number of days of inactivity as a parameter. It then deactivates users who have not logged in since the specified number of days ago.

* Rename Script Includes/InactiveUserCleanup.js to Script Includes/Inactive User/InactiveUserCleanup.js

* Create readme.md

This utility is designed for ServiceNow administrators to help maintain user accounts by automatically deactivating users who have not logged in for a specified number of days. It's a Script Include that can be scheduled to run at regular intervals or executed manually.