Cancel in progress flow executions using flow name (#1200)
* Create script.js

This script can be used to cancel all the in progress flows using the name of the flow you have provided. This helps in cancelling huge number of flows which got created for testing, as well as cancelling huge number of defective executions.

* Create readme.md

readme.md file has been added for the script.js file in "Cancel in progress flow executions using flow name" folder. This file explains how to use the script as well as its use case.