Create displayMonthdays
The provided JavaScript code will display the current month name in the ServiceNow console.

Here's a breakdown of what the code does:

Gets the current date: The currentDate variable is initialized with the current date using the new Date() constructor.
Gets the month index: The getMonth() method of the currentDate object returns the current month's index, which is a number from 0 to 11 (January is 0, February is 1, and so on).
Gets the month name: An array of month names is defined. The monthIndex is used to access the corresponding month name from the array.
Displays the month name: The gs.info() function is used to log the current month name to the console.