CMDB Utility scripts added to improve the effectiveness and precision of CMDB management (#1473)
* Created unUsedCIs.js to detect stale CIs

This script is used to identify unused Configuration Items (CIs) in the CMDB that have not been updated within a specified number of days.

* Create populateMissingManufacturers.js

 This script is used in ServiceNow to automatically populate the manufacturer field for CIs (Configuration Items) in the CMDB (Configuration Management Database) .

* Create detectDuplicateCIs.js

 This script is designed to detect duplicate Configuration Items (CIs) in the CMDB (Configuration Management Database) in ServiceNow.

* Create readme.md

Created readme.md file which contains the detailed description for the three CMDB Utility scripts. Three CMDB utility scripts improve the effectiveness and precision of CMDB management.