Create Prevent_unnecessary_notifications_from_being_sent_out.js
Created a background script to prevent unnecessary notifications from being sent out. It helps in managing the volume of emails being sent so that we do not send the notifications even by mistake. This script is mostly used in dev or uat to avoid any notifications being sent from lower instances.