Add files via upload
This UI Page was developed to streamline the update process of the Install Status field for CMDB (Configuration Management Database) items in ServiceNow. Using Jelly XML scripting, the UI Page fetches details for a selected CI item, allows users to modify the Install Status, and save changes back to the CMDB record.   Client script and Script include als used to fetch  configuration item data and update it to table.