Find Groups without Group Members (#1610)
* Get Groups with no Members.js

The purpose of this script is to identify and list all active user groups that do not have any members. This can be useful for maintaining user group data, identifying orphaned groups, or cleaning up unnecessary entries.

* readme.md

The purpose of this script is to identify and list all active user groups that do not have any members. This can be useful for maintaining user group data, identifying orphaned groups, or cleaning up unnecessary entries.