Auto Disable the Service Accounts due to inactivity (#1646)
* Create Disable Service accounts due to inactivity

This schedule job script will help to auto disable the service accounts which has not been logged in or used for last 40 days.

* Create Readme.md

This file gives the information about the above file for auto disabling the service accounts

* Disable Service accounts due to inactivity

This is for disabling the Service account due to inactivity

* Readme.md

This schedule job script will help to auto disable the service accounts which has not been logged in or used for last 30 days. This can be helpful for those who are looking to disable the accounts based on some certain time period.

* Disable Service accounts due to inactivity

This code helps to disable the service account due to inactivity for last 30 days.