Code snippet - Checking Specific Day and Execute Logic (#1697)
* creation of Do Task on weekdays

Code that verifies today is weekend or not. If today is not weekend, system can trigger mails or execute scripts or do some other updates based on current data.

* Create README.md

Script for checking today is weekday and execute or trigger next part.

* creation of Do Task on weekdays

Check if today is weekday and execute the task to be executed everyday except Saturday and Sunday.

* Rename Do Task on weekdays to Do Task on weekdays.js

* Update Do Task on weekdays.js

* Update README.md

* Rename Do Task on weekdays.js to doTaskonWeekdays.js

* Update README.md