Readme.md
Advance condition section in notification, enables with the capability to filter any kind of data present in the system and based on that you can trigger the notification. In my scenario, I have done it for the specific catalog item, whenever you request any catalog item, this notification would only trogger when the sys_id matches with the particular catalog item's sys_Id.

Note: Advance condition section always returns the Boolean value as same as we have in entitlement scripts and if condition passes it returns true or else it returns false.