Utilizing spModal functionalities for SweetAlerts (#1740)
* spModalSweetAlerts.js

spModal alerts are modern sweet alert to show the choices to be selected in an interactive way.

Here, we are providing HR's to select the type of rewards in the modal page and populate a field with the selected value and show a message of the selected reward.

* readme.md

spModal alerts are modern sweet alert to show the choices to be selected in an interactive way.

Here, we are providing HR's to select the type of rewards in the modal page and populate a field with the selected value and show a message of the selected reward.

* readme.md

spModal alerts are modern sweet alert to show the choices to be selected in an interactive way.

Here, we are providing HR's to select the type of rewards in the modal page and populate a field with the selected value and show a message of the selected reward.