Detect oldValue, newValue and Operation type in Glide List (#1763)
* detectOldValuenewValueOperation.js

This onChange Client script comes handy when dealing with Glide List type fields where we have to detect whether the value was added or removed and return the sysIDs of items which was added/removed.

* readme.md

This onChange Client script comes handy when dealing with Glide List type fields where we have to detect whether the value was added or removed and return the sysIDs of items which was added/removed.

* detectOldValuenewValueOperation.js

This onChange Client script comes handy when dealing with Glide List type fields where we have to detect whether the value was added or removed and returns the display name of users who were added/removed along with name of operation performed.

* readme.md

This onChange Client script comes handy when dealing with Glide List type fields where we have to detect whether the value was added or removed and returns the name of users who were added/removed.

* scriptUtil.js

This script include handles the sysIds(oldValue & newValue) of users and returns the display name

* readme.md

This onChange Client script comes handy when dealing with Glide List type fields where we have to detect whether the value was added or removed and return the display name of users who were added/removed.

* watchListCandidatesUtil.js

This script include handles the sysIds(oldValue & newValue) of users and returns the display name

* detectOldValuenewValueOperation.js

This onChange Client script comes handy when dealing with Glide List type fields where we have to detect whether the value was added or removed and returns the display name of users who were added/removed along with name of operation performed.

* readme.md

This onChange Client script comes handy when dealing with Glide List type fields where we have to detect whether the value was added or removed and returns the display name of users who were added/removed along with name of operation performed.

* watchListCandidatesUtil.js

This script include handles the sysIds(oldValue & newValue) of users and returns the display name