Fetch dynamic value from decision table (#1920)
* Create Task

Creating task when button is clicked with the same details added in Incident ticket.

* Fetch dynamic value from decision table

If we want to have a scenario like based on "FieldA" of big list of comma separated values, We need to provide its relevant "FieldB" value, Which need not to be repeated again and the key value pair is huge and often changing, In that case we can make use of decision table to fetch key value pair without interrupting any scripts and saving value in "FieldB" which will get changed according to "FieldA" value as it is a calculated field.

* README.md

Read me file for calculated field using decision builder

* Delete UI Actions/Create Task

---------

Co-authored-by: Earl Duque <31702109+earlduque@users.noreply.github.com>