Change close reminder business rule (#1933)
* Create change_close_reminder.js

This Business Rule automatically prepares an email reminder for Change Requests that are about to close.
When triggered, it dynamically builds a clear and actionable subject and body message to alert users that the Change Request is nearing closure.

* Create README.md

This Business Rule automatically prepares an email reminder for Change Requests that are about to close.
When triggered, it dynamically builds a clear and actionable subject and body message to alert users that the Change Request is nearing closure.

* Update README.md

This Business Rule automatically generate a reminder message for Tasks that are approaching their closure time.

* Update README.md