Check if Fiscal Year is safe for Deletion
You can do a lot of damage related to data corruption and impacted financials if you delete your Fiscal Periods/Fiscal Year when it's already in use with live data like Cost Plans and Benefit Plans. 

This will validate a single Fiscal Year to validate it's not used yet and safe for deletion if your Fiscal Periods have Validation issues.

Use this in the [System Definition > Scripts - Background] module. Enter in the Fiscal Year by display name on line 2.