Noah drew patch 2 (#1961)
* Create Explain Project EAC (Estimate At Completion) Value

The Estimate At Completion value (EAC) for Projects can be hard to visualize. This script will print out messages to explain the calculation and how the value is getting calculated.

Run this in the [System Defintion > Scripts - Background] module. Set the "projectId" variable on line 1 with the Sys ID of your pm_project record

* Delete Specialized Areas/Project Management/Explain Project EAC (Estimate At Completion) Value directory

* Create Create Explain Project EAC (Estimate At Completion) Value

* Create README.md