Issue 1903/fix naming combine (#1977)
* Update README.md

Updated the readme to combine the readme for both JS files. There are now 2 readme's which is confusing and reported in issue #1903

* Delete Core ServiceNow APIs/GlideRecord/Get Reference Record/Readme_Soumyadeep.md

Removed additional readme, as the naming is not in line, and another readme already exists. Updated the other readme to include the content of this one + improvements.

* Rename script.js to get_assignment_group_from_incident.js

Rename the script to account for the second JS file in this folder. This is the name also referred to in the readme

* Rename GetRecordRequestedBy_Soumyadeep.js to get_requested_by_user.js

Renamed the file to account for a second js in this folder. This file name is also referenced in the updated readme.