Smart Field Validation and Dependent Field Derivation Using getError() and setError() (#2036)
* Create custom_alert.js

If you need HTML beautified alert box, use this.

* Create custom_alert_box.js

UI Page to render the custom alert

* ExampleScreenShots

These are example ScreenShots for reference

* Create readme.md

Example Screenshots

* Rename Catalog Client Script/CustomAlert/ExampleScreenShotCustomAlertInfo.png to Catalog Client Script/CustomAlert/Screenshots/ExampleScreenShotCustomAlertInfo.png

* Rename Catalog Client Script/CustomAlert/ExampleScreenShotForCustomAlertSuccess.png to Catalog Client Script/CustomAlert/Screenshots/ExampleScreenShotForCustomAlertSuccess.png

* Rename Catalog Client Script/CustomAlert/ExampleScrenShotForCustomAlertDanger.png to Catalog Client Script/CustomAlert/Screenshots/ExampleScrenShotForCustomAlertDanger.png

* Create readme.md

Notes for installation or Implementation

* Create update_notes_tag_addition.js

Business rule to add work notes on record if notes for tag entries are allowed.
It is very hard to track who added the tag to the record and when, this will help to manage to understand who has added what tags on record.

* Create update_notes_tag_removal.js

Business rule to add work notes on record if notes for tag removal are allowed.
It is very hard to track who removed the tag to the record and when, this will help to manage to understand who has removed what tags on record.

* Create readme.md

Notes for the implementation

* Revert "Create update_notes_tag_addition.js"

This reverts commit a56fd1fb5cea7887e8a6bca53898febc6c750ced.

* Revert "Create update_notes_tag_removal.js"

This reverts commit deb75b1add391f9f0ea43b81fce6c2b7775d3560.

* Revert "Create readme.md"

This reverts commit 064f91933c35457db6ab487f7b9fcd466c4189c6.

* Create br_validate_short_description.js

* Create README.md

* Create br_derive_dependent_fields.js

* Update br_derive_dependent_fields.js

* Update README.md

* Delete GlideElement/Smart Field Validation and Dependent Field Derivation Using getError() and setError() directory

* Create README.md

Readme File

* Create br_validate_short_description.js

Validate Short Description BR

* Create br_derive_dependent_fields.js

Dependent BR