Fix script pull (#2164)
* Create README.md

This README explains how to configure and use the fix script.

* Create fix_script.js

This Fix Script ensures all users in a specified group have the same roles assigned to that group.
If any user is missing group roles, they are temporarily removed and re-added to the group to refresh their role assignments.